+++ src/tests/TestClearColor.hpp
@@ -9,7 +9,7 @@ namespace test {
 		TestClearColor();
 		~TestClearColor();
 		
-		void onUpdate(float deltaTime) override;
+		void onUpdate(double deltaTime) override;
 		void onRender() override;
 		void onImGuiRender() override;
 	private:
